Pdf_editor/backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pdf2docx
docx2pdf
reportlab
pdfminer.six
pikepdf
Pillow
PyMuPDF
